More accurate i think

+++ include/components/body.hpp
@@ -3,10 +3,10 @@
 #include <string>
 
 struct Body {
-  float mass;
-  float radius;
+  double mass;
+  double radius;
   std::string name;
 
-  Body(float mass, float radius, std::string name)
+  Body(double mass, double radius, std::string name)
       : mass(mass), radius(radius), name(std::move(name)) {}
 };
